==> raleigh/code/22-exploratory-spatial.Rmd <==
---
title: "Exploratory Spatial Analysis"
output:
  html_document: default
  html_notebook: default
---

```{r loadLibraries}
library(readr)
library(data.table)
library(knitr)
library(ggmap)
library(leaflet)
library(RColorBrewer)
```

```{r readData}
dat <- read_csv("../data/merged.csv") %>% data.table()
```

```{r meanCriticalByZip}
dat[ , .(N = .N, 
         mean_num_critical = mean(num_critical), 
         sd_num_critical = sd(num_critical)), 
     by = zip][order(-mean_num_critical)] %>% 
  kable()
```

## Map analysis

Here we use `ggmap`. 

```{r staticMap, fig.width = 8, fig.height = 6}
map <- get_map(location='raleigh', maptype = "terrain", zoom = 10,
               source='google',color='color')
ggmap(map) + geom_point(size = 1,
                        aes(x=X, y=Y, show_guide = TRUE, colour=num_critical), 
                        data=dat, alpha=.5, na.rm = TRUE)  + 
  scale_color_gradient(low="blue", high="red")
```

Analyze inspector patterns. 

```{r staticMapInspector, fig.width = 14, fig.height = 10}
# Take unique locations by inspector, so that we aren't mapping many inspections 
# by same inspector at same location. 
dat_inspector_for_map <- dat[ , list(X, Y), by = InspectedBy] %>% unique()
map <- get_map(location='raleigh', maptype = "terrain", zoom = 11,
               source='google',color='bw')
ggmap(map) + 
  geom_point(size = 2, data=dat_inspector_for_map, na.rm = TRUE,
             aes(x=X, y=Y, colour= InspectedBy))

# Lighter background with hollow circles. 
map <- get_map(location='raleigh', maptype = "toner-lite", zoom = 11,
               source='stamen', color='bw')
ggmap(map) + 
  geom_point(size = 3, data=dat_inspector_for_map, na.rm = TRUE, shape = 1,  # hollow circles
             aes(x=X, y=Y, colour=factor(InspectedBy))) + 
  scale_shape(solid = FALSE)
```

Here we use the `leaflet` package. This one is interactive. 

```{r leafletMap, echo=FALSE, warning=FALSE, message=FALSE, fig.width = 8, fig.height = 6}
dat_map <- subset(dat, select = c("HSISID", "num_critical", "X", "Y"))
setnames(dat_map, c("X", "Y"), c("lng", "lat"))
dat_map <- subset(dat_map, !(lng == 0 | lat == 0))

dat_pal <- colorBin(palette = "Blues", domain = dat_map$num_critical, n = 6)

pal <- brewer.pal(6, "YlOrRd")
dat_pal <- colorBin(palette = pal, domain = dat_map$num_critical, 
                    bins = c(0, 1, 3, 5, 8, 25))

m <- leaflet(dat_map) %>%
  addTiles() %>% 
  addCircles(lng = ~lng, lat = ~lat, color = ~dat_pal(num_critical)) %>% 
  addLegend("bottomright", pal = dat_pal, values = ~num_critical,
            title = "Number of Critical Violations",
            opacity = 1)
m  # print the map
```
